Add axios to devDependencies and enhance pulsar-consumer tests
Added 'axios' to the devDependencies in the package.json file, enabling the use of this tool in development and testing environments. Additionally, made several improvements to the test coverage for the 'pulsar-consumer.js'. This includes better management of the Pulsar container and the addition of a new file: 'pulsar-container.js'.