Update connection status in Pulsar reader
The code for the Pulsar reader has been updated to provide visual feedback following connection status. After successfully creating a reader, it will now show a green dot symbol and the text "connected". In case of a connection error, it will display a red dot and the text "Connection error".