Update dependencies, modify schema logic, and refactor tests
The commit makes several updates to the codebase. It updates the versions of 'pulsar-client' and removes 'ts-log' from 'package-lock.json'. Additionally, it modifies the schema logic in 'pulsar-schema.ts' to change 'parseMandatoryEnum' to 'parseMandatoryChoice' and makes changes to 'Nodes.spec.ts', to improve test module organization and logic. Lastly, it replaces the use of 'v4' with 'uuidv4' for subscription and consumer name generation.