Update Pulsar Producer to accept additional properties
This update includes changes to the Pulsar Producer node, which now takes additional properties from the input message and includes them in the outgoing Pulsar message. This enhancement provides improved control over the Pulsar message properties and includes the addition of several utility conversion functions to facilitate the change. It also updates the corresponding tests and documentation. Minor CHANGELOG version is bumped up as well.